Fix rANS table size estimate for symbols above INT32_MAX
ApproximateRAnsFrequencyTableBits() takes |max_value| as int32_t, but
every caller derives it from a symbol value, which ranges over the whole
of uint32_t:

  ApproximateRAnsFrequencyTableBits(entropy_data.max_symbol + 1, ...)

EntropyData::max_symbol is likewise an int, so a symbol above INT32_MAX
is stored reinterpreted as negative and arrives negative. The estimate
then computes

  8 * (num_unique_symbols + (max_value - num_unique_symbols) / 64)

on a large negative |max_value| and returns a large negative number of
bits where the correct answer is a large positive one.

That result feeds the cost comparison in
MeshPredictionSchemeConstrainedMultiParallelogramEncoder, which picks
the configuration with the fewest estimated bits. A configuration whose
residuals are so bad that they overflow into this region is scored as
cheaper than every reasonable one, so the encoder selects it. The
prediction is still decoded correctly - the choice is recorded in the
stream - but the output is larger than it should be, and which
configuration wins is not the one the search was meant to find.

Taking |max_value| as uint32_t, guarding the subtraction, and widening
before the multiply keeps the estimate monotonic in the symbol value.
Reproduced by encoding a 250-vertex mesh with positions, normals and
texture coordinates at 19/14/29-bit quantization and speed 1: upstream
emits 4524 bytes against the 4546 that the unclamped estimate selects.
Found while diffing Draco's encoder against an independent
implementation of the same bitstream on randomly generated meshes.

+++ src/draco/compression/entropy/shannon_entropy.cc
@@ -140,8 +140,12 @@ int64_t ShannonEntropyTracker::GetNumberOfDataBits(
 
 int64_t ShannonEntropyTracker::GetNumberOfRAnsTableBits(
     const EntropyData &entropy_data) {
-  return ApproximateRAnsFrequencyTableBits(entropy_data.max_symbol + 1,
-                                           entropy_data.num_unique_symbols);
+  // |max_symbol| is an int but holds a symbol value, so a symbol above
+  // INT32_MAX is stored in it reinterpreted as negative. Casting back to
+  // uint32_t recovers the original magnitude for the estimate.
+  return ApproximateRAnsFrequencyTableBits(
+      static_cast<uint32_t>(entropy_data.max_symbol + 1),
+      entropy_data.num_unique_symbols);
 }
 
 }  // namespace draco
